docker-compose spins up 5 api instances now
i know that docker-compose can start more containers but i had to use
specific ports for using nginx in a standalone way and not in a
container. so i defined all containers in the docker-compose.yml and
make the later containers depend on the first api image that gets build.
its important to run `docker-compose build` before `docker-compose up`
before `docker-compose up`. else `up` fails because the image is not
there.